+++ src/Components/Form/Radio.php
@@ -44,7 +44,6 @@ class Radio extends VitrineComponent
 
     protected static array $assets = [
         'js' => ['behaviors/Input.js'],
-        'css' => ['components/form/radio.css', 'components/form/input.css'],
     ];
 
     public function __construct(